Add files via upload
Welcome to the Terminal-based Rock, Paper, Scissors game!

Experience the thrill of this classic game in an innovative way, with an interactive user interface that allows you to select your choice and see the results on the screen.

Our game utilizes pyFIGlet and Prettytable to enhance the visual experience and make the game even more fun. You won't be able to resist playing again and again!

Are you ready to show off your Rock, Paper, Scissors skills? Download our game and start playing today!